+++ RELEASE_NOTES.md
@@ -8,6 +8,7 @@ A new header is inserted each time a *tag* is created.
 - Fixed morph shapes not rendering in WebGL.
 - Added support for the latest version of emscripten.
 - gltfio: fixed blackness seen with default material.
+- Rendering a SAMPLER_EXTERNAL texture before setting an external image no longer crashes the Metal backend.
 
 ## v1.4.2